Library

Contents

Library

Contents

Public

Private

Index

Public

Documentation for TestReports.jl's public interface.

TestReports.test — Function

TestReports.test(; kwargs...)
TestReports.test(pkg::Union{AbstractString, Vector{AbstractString}; kwargs...)

Keyword arguments:

coverage::Bool=false: enable or disable generation of coverage statistics.

julia_args::Union{Cmd, Vector{String}}: options to be passed to the test process.

test_args::Union{Cmd, Vector{String}}: test arguments (ARGS) available in the test process.

logfilepath::AbstractString=pwd(): file path where test reports are saved.

logfilename::Union{AbstractString, Vector{AbstractString}}: name(s) of test report file(s).

Generates a JUnit XML for the tests of package pkg, or for the current project (which thus needs to be a package) if no positional argument is given to TestReports.test. The test report is saved in the current working directory and called testlog.xml if both logfilepath and logfilename are not supplied. If pkg is of type Vector{String}, the report filenames are prepended with the package name, for example Example_testlog.xml.

If logfilename is supplied, it must match the type (and length, if a vector) of pkg.

The tests are run in the same way as Pkg.test.

TestReports.recordproperty — Function

recordproperty(name::String, value)

Adds a property to a testset with name and value that will in turn be added to the <properties> node of the corresponding testsuite in the JUnit XML.

Multiple properties can be added to one testset, but if the same property is set on both parent and child testsets, the value in the child testset takes precedence over that in the parent.

The suggested use of this function is to place it inside a testset with unspecified type (see Examples). This will ensure that Pkg.test is unnaffected, but that the properties are added to the report when TestReports.test is used. This is because properties are only added when the Testset type has a TestReports.properties method defined, as does the ReportingTestSet used by TestReports. TestReports.properties can be extended for custom TestSets.

If a child testset has this method defined but its parent doesn't, the property should be in the report when TestReport.test is used, assuming that the parent testset type doesn't do anything to affect the reporting behaviour. However this is not tested functionality.

value must be serializable by EzML, which gives quite a lot of freedom.

Examples

using TestReports

# Default testset used, so function will not affect Pkg.test but will be used when
# generating JUnit XML.
@testset "MyTestSet" begin
    recordproperty("ID", 42)
    recordproperty("File", @__FILE__)
    recordproperty("Bool", true)
    @test 1==1
    @test 2==2
end

See also: properties

TestReports.ReportingTestSet — Type

ReportingTestSet

Custom AbstractTestSet type designed to be used by TestReports.jl for creation of JUnit XMLs.

Does not throw an error when a test fails or has an error. Upon finishing, a ReportingTestSet will display the default test output, and then flatten to a structure that is suitable for report generation.

It is designed to be wrapped around a package's runtests.jl file and this is assumed when both the test results are displayed and when the TestSet is flatted upon finish. See bin/reporttests.jl for an example of this use. ReportingTestSets are not designed to be used directly in a package's tests, and this is not recommended or supported.

A ReportingTestSet has the description and results fields as per a DefaultTestSet, and has the following additional fields:

properties: a dictionary which is used to record properties to be inserted into the report.

start_time::DateTime: the start date and time of the testing (local system time).

time_taken::Millisecond: the time taken in milliseconds to run the TestSet.

last_record_time::DateTime: the time when record was last called.

hostname::String: the name of host on which the testset was executed.

See also: flatten_results!, recordproperty, report

TestReports.any_problems — Function

any_problems(ts)

Checks a testset to see if there were any problems (Errors or Fails). Note that unlike the DefaultTestSet, the ReportingTestSet does not throw an exception on a failure. Thus to set the exit code from the runner code, we check it using exit(any_problems(top_level_testset)).

TestReports.report — Function

report(ts::AbstractTestSet)

Will produce an XMLDocument that contains a report about the TestSet's results. This report will follow the JUnit XML schema.

To report correctly, the TestSet must have the following structure:

AbstractTestSet
  └─ AbstractTestSet
       └─ AbstractResult

That is, the results of the top level TestSet must all be AbstractTestSets, and the results of those TestSets must all be Results.

Additionally, all of the AbstractTestSets must have both description and results fields.

Private

Package internals documentation.

Report Generation

TestReports.TESTS_FAILED — Constant

Exit code for runner when tests fail

TestReports.checkexitcode! — Method

checkexitcode!(errs, proc, pkg, logfilename)

Checks proc.exitcode and acts as follows:

If 0, displays tests passed info message

If equal to TESTS_FAILED const, warning is displayed and pkg added to errs

If anything else, throws a PkgTestError

TestReports.gen_command — Method

gen_command(runner_code, julia_args, coverage)

Returns Cmd which will run the runner code in a new Julia instance.

See also: gen_runner_code

TestReports.gen_runner_code — Method

gen_runner_code(testfilename, logfilename, test_args)

Returns runner code that will run the tests and generate the report in a new Julia instance.

TestReports.get_deps! — Method

get_deps!(deps, manifest, pkg)

Push dependencies for pkg found in manifest into deps.

TestReports.get_deps — Method

get_deps(manifest, pkg) = get_deps!(String[], manifest, pkg)

Get list of dependencies for pkg found in manifest

TestReports.get_manifest — Method

get_manifest()

Return the parsed manifest that has TestReports as a dependency.

TestReports.get_testreports_environment — Method

get_testreports_environment()

Returns new environment to be pushed to LOAD_PATH to ensure TestReports, Test and their dependencies are available for report generation.

TestReports.gettestfilepath — Method

gettestfilepath(ctx::Context, pkgspec::Pkg.Types.PackageSpec)

Gets the testfile path of the package. Code for each Julia version mirrors that found in Pkg/src/Operations.jl.

TestReports.isinstalled! — Method

isinstalled!(ctx::Context, pkgspec::Pkg.Types.PackageSpec)

Checks if the package is installed by using ensure_resolved from Pkg/src/Types.jl. This function fails if the package is not installed, but here we wrap it in a try-catch as we may want to test another package after the one that isn't installed.

For Julia versions V1.4 and later, the first arguments of the Pkg functions used is of type Pkg.Types.Context. For earlier versions, they are of type Pkg.Types.EnvCache.

TestReports.make_testreports_environment — Method

make_testreports_environment(manifest)

Make a new environment in a temporary directory, using information from the parsed manifest provided.

TestReports.runtests! — Method

runtests!(errs::Vector, pkg, cmd, logfilename)

Runs cmd which will run the tests of pkg. The exit code of the process is then checked.

TestReports.test! — Method

test!(pkg::AbstractString,
      errs::Vector{AbstractString},
      nopkgs::Vector{AbstractString},
      notests::Vector{AbstractString},
      logfilename::AbstractString;
      coverage::Bool=false,
      julia_args::Union{Cmd, AbstractVector{<:AbstractString}}=``,
      test_args::Union{Cmd, AbstractVector{<:AbstractString}}=``)

Tests pkg and save report to logfilename. Tests are run in the same way as Pkg.test.

If tests error pkg is added to nopkgs. If pkg has no testfile it is added to notests. If pkg is not installed it is added to nopkgs.

TestSets

TestReports.ReportingResult — Type

ReportingResult{T}

Wraps a Result of type T so that additional metadata can be saved.

fields

result::T: Result that is being wrapped.

time_taken::Millisecond: the time taken (milliseconds) for this test to be run.

TestReports._flatten_results! — Method

_flatten_results!(ts::AbstractTestSet)::Vector{<:AbstractTestSet}

Recursively flatten ts to a vector of TestSets.

TestReports._flatten_results! — Method

_flatten_results!(rs::Result)

Return vector containing rs so that when iterated through, rs is added to the results vector.

TestReports.add_to_ts_default! — Method

add_to_ts_default!(ts_default::DefaultTestSet, result::Result)
add_to_ts_default!(ts_default::DefaultTestSet, result::ReportingResult)
add_to_ts_default!(ts_default::DefaultTestSet, ts::AbstractTestSet)
add_to_ts_default!(ts_default::DefaultTestSet, ts::ReportingTestSet)

Populate ts_default with the supplied variable. If result is a Result or an AbstractTestSet (but not a ReportingTestSet) then it is recorded. If it is a ReportingTestSet then a new DefaultTestSet with matching description is created, populated by recursively calling this function and then added to the results of ts_default. If result is a ReportingResult, the Result contained by the ReportingResult is added to ts_default.

TestReports.display_reporting_testset — Method

display_reporting_testset(ts::ReportingTestSet)

Displays the test output in the same format as Pkg.test by using a DefaultTestSet.

TestReports.flatten_results! — Method

flatten_results!(ts::AbstractTestSet)

Returns a flat structure 3 deep, of TestSet -> TestSet -> Result. This is necessary for writing a report, as a JUnit XML does not allow one testsuite to be nested in another. The top level TestSet becomes the testsuites element, and the middle level TestSets become individual testsuite elements, and the Results become the testcase elements.

If ts.results contains any Results, these are added into a new TestSet with the description "Top level tests", which then replaces them in ts.results.

TestReports.handle_top_level_results! — Method

handle_top_level_results!(ts::AbstractTestSet)

If ts.results contains any Results, these are removed from ts.results and added to a new ReportingTestSet, which in turn is added to ts.results. This leaves ts.results only containing AbstractTestSets.

The time_taken field of the new ReportingTestSet is calculated by summing the time taken by the individual results, and the start_time field is set to the start_time field of ts.

TestReports.hostname — Method

hostname(ts::ReportingTestSet)
hostname(ts::AbstractTestSet)

Get the hostname of a ReportingTestSet, returns gethostname() for an AbstractTestSet. Can be extended for custom TestSets, must return a string.

TestReports.ispass — Method

ispass(result)

Return true if result is a Pass or a ReportingResult{Pass}, otherwise return false.

TestReports.properties — Method

properties(ts::ReportingTestSet)
properties(ts::AbstractTestSet)

Get the properties dictionary of a ReportingTestSet, returns nothing for an AbstractTestSet. Can be extended for custom TestSets, and must return either a Dict or nothing.

TestReports.set_start_time! — Method

set_start_time!(ts::ReportingTestSet, start_time)
set_start_time!(ts::AbstractTestSet, start_time)

Sets the start time field of a ReportingTestSet. This is used when flattening ReportingTestSets for report generation and an be extended for custom TestSets.

TestReports.set_time_taken! — Method

set_time_taken!(ts::ReportingTestSet, time_taken)
set_time_taken!(ts::AbstractTestSet, time_taken)

Sets the time taken field of a ReportingTestSet. This is used when flattening ReportingTestSets for report generation and an be extended for custom TestSets.

TestReports.start_time — Method

start_time(ts::ReportingTestSet)
start_time(ts::AbstractTestSet)

Get the start time of a ReportingTestSet, returns Dates.now() for an AbstractTestSet. Can be extended for custom TestSets, must return a DateTime.

TestReports.time_taken — Method

time_taken(ts::ReportingTestSet)
time_taken(ts::AbstractTestSet)

Get the time taken of a ReportingTestSet, returns Dates.Millisecond(0) for an AbstractTestSet. Can be extended for custom TestSets, must return a Dates.Millisecond.

TestReports.time_taken — Method

time_taken(result::ReportingResult)
time_taken(result::Result)

For a ReportingResult, return the time taken for the test to run. For a Result, return Dates.Millisecond(0).

TestReports.update_testset_properties! — Method

update_testset_properties!(childts::AbstractTestSet, ts::AbstractTestSet)

Adds properties of ts to childts. If any properties being added already exist in childts, a warning is displayed and the value in ts is overwritten.

If the types of ts and\or childts do not a method defined for TestReports.properties, this is handled as follows:

If method not defined for typeof(ts), it has no properties to add to childts and therefore nothing happens.

If method not defined for typeof(chidlts) and ts has properties, then a warning is shown.

See also: properties

XML Writing

TestReports.add_testsuite_properties! — Method

add_testsuite_properties!(x_testsuite, ts::AbstractTestSet)

Add all key value pairs in the properties field of a AbstractTestSet to the corresponding testsuite xml element. This function assumes that the type of ts has a TestReports.properties method defined.

See also: properties

TestReports.check_ts — Method

check_ts(ts::AbstractTestSet)

Throws an exception if ts does not have the right structure for report or if the results of ts do not have both description or results fields.

See also: report

TestReports.error_xml — Method

failure_xml(message, test_type, content)

Create an error node (which will be the child of a testcase).

TestReports.failure_xml — Method

failure_xml(message, test_type, content)

Create a failure node (which will be the child of a testcase).

TestReports.format_period_for_xml — Method

formatperiodfor_xml(time::Millisecond)

Formats a millisecond value into a string in seconds with 3 decimal places.

@sprintf is used to ensure that value does not use scientific notification, which is not allowed in an XML decimal.

Methods for other Periods have not been written as are not currently required.

TestReports.get_error_info — Method

get_error_info(v::Error)

Return message and type of error for testcase attribute, and number of tests (either 1 or 0). Uses test_type field to determine what caused the original error.

TestReports.get_failure_message — Method

getfailuremessage(v::Fail)

Return message for failed test testcase attribute. Uses test_type field to determine what caused the original failure.

TestReports.set_attribute! — Method

set_attribute!(node, attr, val)
set_attribute!(node, attr, val::Period)

Add the attritube with name attr and value val to node.

If val is of type Period, format with format_period_for_xml.

TestReports.skip_xml — Method

skip_xml()

Create a skip node (which will be the child of a testcase).

TestReports.testcase_xml — Method

testcase_xml(name, id, x_children)

Create a testcase element of a JUnit XML.

This is the generic form (with name, id and children) that is used by other methods.

TestReports.testcase_xml — Method

testcase_xml(v::Result, childs)

Create a testcase element of a JUnit XML for the result v.

The original expression of the test is used as the name, whilst the id is defaulted to testcaseid_.

TestReports.testsuite_xml — Method

testsuite_xml(name, ntests, nfails, nerrors, x_children)

Create a testsuite element of a JUnit XML.

TestReports.testsuites_xml — Method

testsuites_xml(ntests, nfails, nerrors, x_children)

Create the testsuites element of a JUnit XML.

TestReports.to_xml — Method

to_xml(ts::AbstractTestSet)

Create a testsuite node from a AbstractTestSet, by creating nodes for each result in ts.results. For creating a JUnit XML, all results must be AbstractResults, that is they cannot be AbstractTestSets, as the XML cannot have one testsuite nested inside another.

TestReports.to_xml — Method

to_xml(res::Pass)
to_xml(res::Fail)
to_xml(res::Broken)
to_xml(res::Error)
to_xml(res::ReportingResult)

Create a testcase node from the result and return it along with information on number of tests.

Index

TestReports.TESTS_FAILED

TestReports.ReportingResult

TestReports.ReportingTestSet

TestReports._flatten_results!

TestReports._flatten_results!

TestReports.add_testsuite_properties!

TestReports.add_to_ts_default!

TestReports.any_problems

TestReports.check_ts

TestReports.checkexitcode!

TestReports.display_reporting_testset

TestReports.error_xml

TestReports.failure_xml

TestReports.flatten_results!

TestReports.format_period_for_xml

TestReports.gen_command

TestReports.gen_runner_code

TestReports.get_deps

TestReports.get_deps!

TestReports.get_error_info

TestReports.get_failure_message

TestReports.get_manifest

TestReports.get_testreports_environment

TestReports.gettestfilepath

TestReports.handle_top_level_results!

TestReports.hostname

TestReports.isinstalled!

TestReports.ispass

TestReports.make_testreports_environment

TestReports.properties

TestReports.recordproperty

TestReports.report

TestReports.runtests!

TestReports.set_attribute!

TestReports.set_start_time!

TestReports.set_time_taken!

TestReports.skip_xml

TestReports.start_time

TestReports.test

TestReports.test!

TestReports.testcase_xml

TestReports.testcase_xml

TestReports.testsuite_xml

TestReports.testsuites_xml

TestReports.time_taken

TestReports.time_taken

TestReports.to_xml

TestReports.to_xml

TestReports.update_testset_properties!